requirements.txt
absl-py==2.1.0
aiohttp==3.9.0
aiosignal==1.3.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
bitarray==2.8.3
blessed==1.20.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.5
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
# Editable install with no version control (fairseq==1.0.0a0)
-e fairseq_ours
fast-bss-eval==0.1.4
filelock==3.9.0
fonttools==4.45.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gpustat==1.1.1
grpcio==1.63.0
huggingface-hub==0.19.4
hydra-core==1.0.7
idna==3.4
importlib-resources==6.1.1
importlib_metadata==7.1.0
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2
lightning-utilities==0.10.0
lit==15.0.7
llvmlite==0.42.0
lxml==4.9.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
networkx==3.0
numba==0.59.1
numpy==1.23.5
nvidia-ml-py==12.535.133
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==9.3.0
platformdirs==4.2.1
pooch==1.8.1
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==1.6.3
pytorch-ranger==0.1.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
rich==13.7.0
sacrebleu==2.3.2
safetensors==0.4.0
scikit-learn==1.4.2
scipy==1.11.4
sentry-sdk==1.36.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soundfile>=0.11.0
soxr==0.3.7
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==0.9.11
tomli==2.0.1
torch-mir-eval==0.4
torch-optimizer==0.3.0
torchmetrics==1.2.0
tqdm==4.66.1
traitlets==5.14.1
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.16.0
wcwidth==0.2.12
Werkzeug==3.0.2
yarl==1.9.3
zipp==3.17.0
npy_append_array==0.9.16
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
# Pin for torch 1.13.x: transformers 4.41+ may require PyTorch>=2.1 (see README).
transformers==4.36.2